tests: Fix tests calling nonexistent or old binary build
